added provider and moved code out of home
Home should be clean. Provider is a bit wrong because it not only manages the state of a single screen. It is more like a global state. I just want to use a presenter... It is definitely not a good idea to put everything inside the page. I will check how others do it tomorrow.